Consolidate battle scripts
