+++ preplan_editor.html
@@ -42,7 +42,10 @@ Developer: Deathsgift66
   <link href="/CSS/resource_bar.css" rel="stylesheet" />
 
   <!-- Scripts -->
-  <script src="/Javascript/preplan_editor.js" type="module"></script>
+  <script type="module">
+    import { initPreplanEditor } from "/Javascript/battle_core.js";
+    initPreplanEditor();
+  </script>
 
 <!-- ✅ Injected standard Thronestead modules -->
   <script src="/Javascript/components/authGuard.js" type="module"></script>
